+++ users.go
@@ -1,11 +1,11 @@
 package main
 
 type userData struct {
-	Id             string
-	Name           string
-	Password       string
-	Email          string
-	Address        string
-	DateOfBirth    string
-	DateOfCreation string
+	Id          string
+	Name        string
+	Password    string
+	Email       string
+	Address     string
+	DateOfBirth string
+	DateOfLogin string
 }
